github-runner: always restart the runner
We may have a race condition between the agenix job and the runner.
If the key isn't available yet, the runner fails to start and isn't
restarted.

Ideally, we'd wait for the key, but launchd doesn't have that
functionality. We'd need to modify the command itself to add a
wait4path.